backend/requirements.txt
django
psycopg2-binary
celery
redis
djangorestframework
requests
pylint
